+++ ali/common/src/main/java/com/yanny/ali/plugin/common/NodeUtils.java
@@ -7,7 +7,6 @@ import com.yanny.ali.api.*;
 import com.yanny.ali.language.Lang;
 import com.yanny.ali.plugin.common.nodes.*;
 import com.yanny.ali.plugin.server.EnchantedRanges;
-import com.yanny.ali.plugin.server.EntryTooltipUtils;
 import com.yanny.ali.plugin.server.TooltipUtils;
 import net.minecraft.core.Holder;
 import net.minecraft.core.Registry;
@@ -50,7 +49,7 @@ public class NodeUtils {
         float chance = getChance(entry, rawChance, sumWeight);
         EnchantedRanges enchantedChance = getEnchantedChance(utils, allConditions, chance);
         EnchantedRanges enchantedCount = getEnchantedCount(utils, allFunctions);
-        TooltipNode tooltip = EntryTooltipUtils.getTooltip(utils, entry.quality, enchantedChance, enchantedCount, allFunctions, allConditions).build();
+        TooltipNode tooltip = TooltipUtils.getTooltip(utils, entry.quality, enchantedChance, enchantedCount, allFunctions, allConditions).build();
         Either<ItemStack, TagKey<? extends ItemLike>> either = itemGetter.apply(allFunctions);
 
         if (either.left().isPresent() && either.left().get().isEmpty()) {
@@ -64,7 +63,7 @@ public class NodeUtils {
     public static AlternativesNode getAlternativesNode(IServerUtils utils, AlternativesEntry entry, float rawChance, int sumWeight, List<LootItemFunction> functions, List<LootItemCondition> conditions) {
         List<LootItemCondition> allConditions = getAllConditions(entry, conditions);
         List<IDataNode> children = getChildren(utils, entry.children, rawChance, sumWeight, functions, allConditions);
-        TooltipNode tooltip = EntryTooltipUtils.getAlternativesTooltip().build();
+        TooltipNode tooltip = TooltipUtils.getAlternativesTooltip().build();
 
         return new AlternativesNode(children, tooltip);
     }
@@ -74,7 +73,7 @@ public class NodeUtils {
         List<LootItemFunction> allFunctions = getAllFunctions(entry, functions);
         List<LootItemCondition> allConditions = getAllConditions(entry, conditions);
         float chance = getChance(entry, rawChance, sumWeight);
-        TooltipNode tooltip = EntryTooltipUtils.getDynamicTooltip(utils, entry.quality, chance, allFunctions, allConditions).build();
+        TooltipNode tooltip = TooltipUtils.getDynamicTooltip(utils, entry.quality, chance, allFunctions, allConditions).build();
 
         return new DynamicNode(chance, tooltip);
     }
@@ -85,7 +84,7 @@ public class NodeUtils {
         List<LootItemCondition> allConditions = getAllConditions(entry, conditions);
         float chance = getChance(entry, rawChance, sumWeight);
         EnchantedRanges enchantedChance = getEnchantedChance(utils, allConditions, chance);
-        TooltipNode tooltip = EntryTooltipUtils.getEmptyTooltip(utils, entry.quality, enchantedChance, allFunctions, allConditions).build();
+        TooltipNode tooltip = TooltipUtils.getEmptyTooltip(utils, entry.quality, enchantedChance, allFunctions, allConditions).build();
 
         return new EmptyNode(chance, tooltip);
     }
@@ -94,7 +93,7 @@ public class NodeUtils {
     public static GroupNode getGroupNode(IServerUtils utils, EntryGroup entry, float rawChance, int sumWeight, List<LootItemFunction> functions, List<LootItemCondition> conditions) {
         List<LootItemCondition> allConditions = getAllConditions(entry, conditions);
         List<IDataNode> children = getChildren(utils, entry.children, rawChance, sumWeight, functions, allConditions);
-        TooltipNode tooltip = EntryTooltipUtils.getGroupTooltip().build();
+        TooltipNode tooltip = TooltipUtils.getGroupTooltip().build();
 
         return new GroupNode(children, tooltip);
     }
@@ -103,7 +102,7 @@ public class NodeUtils {
     public static SequenceNode getSequenceNode(IServerUtils utils, SequentialEntry entry, float rawChance, int sumWeight, List<LootItemFunction> functions, List<LootItemCondition> conditions) {
         List<LootItemCondition> allConditions = getAllConditions(entry, conditions);
         List<IDataNode> children = getChildren(utils, entry.children, rawChance, sumWeight, functions, allConditions);
-        TooltipNode tooltip = EntryTooltipUtils.getSequentialTooltip().build();
+        TooltipNode tooltip = TooltipUtils.getSequentialTooltip().build();
 
         return new SequenceNode(children, tooltip);
     }
@@ -114,7 +113,7 @@ public class NodeUtils {
         List<LootItemCondition> allConditions = getAllConditions(entry, conditions);
         float chance = getChance(entry, rawChance, sumWeight);
         LootTable lootTable = utils.getLootTable(entry.name);
-        TooltipNode tooltip = EntryTooltipUtils.getReferenceTooltip(entry, rawChance, sumWeight).build();
+        TooltipNode tooltip = TooltipUtils.getReferenceTooltip(entry, rawChance, sumWeight).build();
         List<IDataNode> children;
 
         if (lootTable != null) {
@@ -145,7 +144,7 @@ public class NodeUtils {
         List<LootItemFunction> allFunctions = Stream.concat(functions.stream(), Arrays.stream(entry.functions)).toList();
         List<LootItemCondition> allConditions = Stream.concat(conditions.stream(), Arrays.stream(entry.conditions)).toList();
         int sumWeight = getTotalWeight(Arrays.asList(entry.entries));
-        TooltipNode tooltip = EntryTooltipUtils.getLootPoolTooltip(utils.convertNumber(utils, entry.rolls), utils.convertNumber(utils, entry.bonusRolls)).build();
+        TooltipNode tooltip = TooltipUtils.getLootPoolTooltip(utils.convertNumber(utils, entry.rolls), utils.convertNumber(utils, entry.bonusRolls)).build();
         List<IDataNode> children = getChildren(utils, entry.entries, rawChance, sumWeight, allFunctions, allConditions);
 
         return new LootPoolNode(children, tooltip);
@@ -153,7 +152,7 @@ public class NodeUtils {
 
     @NotNull
     public static LootTableNode getLootTableNode(List<ILootModifier<?>> modifiers) {
-        TooltipNode tooltip = EntryTooltipUtils.getLootTableTooltip().build();
+        TooltipNode tooltip = TooltipUtils.getLootTableTooltip().build();
         List<IDataNode> children = new ArrayList<>();
         LootTableNode node = new LootTableNode(children, tooltip);
 
@@ -167,7 +166,7 @@ public class NodeUtils {
     @NotNull
     public static LootTableNode getLootTableNode(List<ILootModifier<?>> modifiers, IServerUtils utils, LootTable entry, float rawChance, List<LootItemFunction> functions, List<LootItemCondition> conditions) {
         List<LootItemFunction> allFunctions = Stream.concat(functions.stream(), Arrays.stream(entry.functions)).toList();
-        TooltipNode tooltip = EntryTooltipUtils.getLootTableTooltip().build();
+        TooltipNode tooltip = TooltipUtils.getLootTableTooltip().build();
         List<IDataNode> children = utils.getLootPools(entry).stream().map((lootPool) -> (IDataNode) getLootPoolNode(utils, lootPool, rawChance, allFunctions, conditions)).toList();
         LootTableNode node = new LootTableNode(children, tooltip);