+++ ali/common/src/main/java/com/yanny/ali/plugin/server/EntryTooltipUtils.java
@@ -1,141 +1,78 @@
 package com.yanny.ali.plugin.server;
 
-import com.yanny.aci.api.RangeValue;
 import com.yanny.aci.tooltip.TooltipBuilder;
 import com.yanny.ali.api.IServerUtils;
 import com.yanny.ali.language.Lang;
-import net.minecraft.world.level.storage.loot.entries.LootPoolSingletonContainer;
-import net.minecraft.world.level.storage.loot.entries.LootTableReference;
-import net.minecraft.world.level.storage.loot.functions.LootItemFunction;
-import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
+import net.minecraft.world.level.storage.loot.entries.*;
 import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
-
 
 public class EntryTooltipUtils {
     @NotNull
-    public static TooltipBuilder getLootTableTooltip() {
-        return TooltipBuilder.keyOnly(Lang.Group.ALL);
-    }
-
-    @NotNull
-    public static TooltipBuilder getReferenceTooltip(LootTableReference entry, float chance, int sumWeight) {
+    public static TooltipBuilder getItemTooltip(IServerUtils utils, LootItem entry) {
         return TooltipBuilder.array((b) -> {
-            b.add(TooltipBuilder.keyOnly(Lang.Group.ALL));
-            b.add(getQualityTooltip(entry.quality));
-            b.add(getChanceTooltip(new EnchantedRanges((chance * entry.weight / sumWeight) * 100)));
-        });
+            b.add(utils.getValueTooltip(utils, entry.item).build(Lang.Value.ITEM));
+            addSingleton(b, utils, entry);
+        }, Lang.Entry.ITEM);
     }
 
     @NotNull
-    public static TooltipBuilder getLootPoolTooltip(RangeValue rolls, RangeValue bonusRolls) {
+    public static TooltipBuilder getTagTooltip(IServerUtils utils, TagEntry entry) {
         return TooltipBuilder.array((b) -> {
-            b.add(TooltipBuilder.keyOnly(Lang.Group.RANDOM));
-            b.add(getRolls(rolls, bonusRolls));
-        });
+            b.add(utils.getValueTooltip(utils, entry.tag).build(Lang.Value.TAG));
+            b.add(utils.getValueTooltip(utils, entry.expand).build(Lang.Value.EXPAND));
+            addSingleton(b, utils, entry);
+        }, Lang.Entry.TAG);
     }
 
     @NotNull
-    public static TooltipBuilder getAlternativesTooltip() {
-        return TooltipBuilder.keyOnly(Lang.Group.ALTERNATIVES);
+    public static TooltipBuilder getEmptyTooltip(IServerUtils utils, EmptyLootItem entry) {
+        return TooltipBuilder.array((b) -> addSingleton(b, utils, entry), Lang.Entry.EMPTY);
     }
 
     @NotNull
-    public static TooltipBuilder getDynamicTooltip(IServerUtils utils, int quality, float chance, List<LootItemFunction> functions, List<LootItemCondition> conditions) {
+    public static TooltipBuilder getDynamicTooltip(IServerUtils utils, DynamicLoot entry) {
         return TooltipBuilder.array((b) -> {
-            b.add(TooltipBuilder.keyOnly(Lang.Group.DYNAMIC));
-            b.add(getQualityTooltip(quality));
-            b.add(getChanceTooltip(new EnchantedRanges(chance * 100)));
-            b.add(GenericTooltipUtils.getConditionsSectionTooltip(utils, conditions));
-            b.add(GenericTooltipUtils.getFunctionsSectionTooltip(utils, functions));
-        });
+            b.add(utils.getValueTooltip(utils, entry.name).build(Lang.Value.NAME));
+            addSingleton(b, utils, entry);
+        }, Lang.Entry.DYNAMIC);
     }
 
     @NotNull
-    public static TooltipBuilder getGroupTooltip() {
-        return TooltipBuilder.keyOnly(Lang.Group.ALL);
-    }
-
-    @NotNull
-    public static TooltipBuilder getSequentialTooltip() {
-        return TooltipBuilder.keyOnly(Lang.Group.SEQUENCE);
-    }
-
-    @NotNull
-    public static TooltipBuilder getEmptyTooltip(IServerUtils utils, int quality, EnchantedRanges chance, List<LootItemFunction> functions, List<LootItemCondition> conditions) {
+    public static TooltipBuilder getReferenceTooltip(IServerUtils utils, LootTableReference entry) {
         return TooltipBuilder.array((b) -> {
-            b.add(TooltipBuilder.keyOnly(Lang.Group.EMPTY));
-            b.add(getQualityTooltip(quality));
-            b.add(getChanceTooltip(chance));
-            b.add(GenericTooltipUtils.getConditionsSectionTooltip(utils, conditions));
-            b.add(GenericTooltipUtils.getFunctionsSectionTooltip(utils, functions));
-        });
+            b.add(utils.getValueTooltip(utils, entry.name).build(Lang.Value.LOOT_TABLE));
+            addSingleton(b, utils, entry);
+        }, Lang.Entry.LOOT_TABLE);
     }
 
     @NotNull
-    public static TooltipBuilder getTooltip(IServerUtils utils, int quality, EnchantedRanges chance, EnchantedRanges count,
-                                               List<LootItemFunction> functions, List<LootItemCondition> conditions) {
-        return TooltipBuilder.array((b) -> {
-            b.add(getQualityTooltip(quality));
-            b.add(getChanceTooltip(chance));
-            b.add(getCountTooltip(count));
-            b.add(GenericTooltipUtils.getConditionsSectionTooltip(utils, conditions));
-            b.add(GenericTooltipUtils.getFunctionsSectionTooltip(utils, functions));
-        });
+    public static TooltipBuilder getAlternativesTooltip(IServerUtils utils, AlternativesEntry entry) {
+        return TooltipBuilder.array((b) -> addComposite(b, utils, entry), Lang.Entry.ALTERNATIVES);
     }
 
     @NotNull
-    public static TooltipBuilder getQualityTooltip(int quality) {
-        if (quality != LootPoolSingletonContainer.DEFAULT_QUALITY) {
-            return TooltipBuilder.value(quality).key(Lang.Description.QUALITY);
-        }
-
-        return TooltipBuilder.empty();
+    public static TooltipBuilder getGroupTooltip(IServerUtils utils, EntryGroup entry) {
+        return TooltipBuilder.array((b) -> addComposite(b, utils, entry), Lang.Entry.GROUP);
     }
 
     @NotNull
-    public static TooltipBuilder getChanceTooltip(EnchantedRanges chance) {
-        RangeValue defaultChance = chance.getUnenchantedValue();
-
-        if (!defaultChance.isRange() && defaultChance.max() > 99.99999) {
-            return TooltipBuilder.empty();
-        }
-
-        TooltipBuilder builder = TooltipBuilder.value(defaultChance, "%");
-
-        chance.forEachEnchantment((enchantment, level, value) -> builder.add(TooltipBuilder.value(
-                value + "%",
-                TooltipBuilder.translate(enchantment.getDescriptionId()),
-                TooltipBuilder.translate("enchantment.level." + level)
-        ).build(Lang.Description.CHANCE_BONUS)));
-
-        return builder.key(Lang.Description.CHANCE);
+    public static TooltipBuilder getSequentialTooltip(IServerUtils utils, SequentialEntry entry) {
+        return TooltipBuilder.array((b) -> addComposite(b, utils, entry), Lang.Entry.SEQUENCE);
     }
 
-    @NotNull
-    public static TooltipBuilder getCountTooltip(EnchantedRanges count) {
-        TooltipBuilder builder = TooltipBuilder.value(count.getUnenchantedValue());
-
-        count.forEachEnchantment((enchantment, level, value) -> builder.add(TooltipBuilder.value(
-                value,
-                TooltipBuilder.translate(enchantment.getDescriptionId()),
-                TooltipBuilder.translate("enchantment.level." + level)
-        ).build(Lang.Description.COUNT_BONUS)));
-
-        return builder.key(Lang.Description.COUNT);
+    private static void addSingleton(TooltipBuilder builder, IServerUtils utils, LootPoolSingletonContainer entry) {
+        builder.add(TooltipUtils.getWeightTooltip(entry.weight));
+        builder.add(TooltipUtils.getQualityTooltip(entry.quality));
+        builder.add(utils.getValueTooltip(utils, entry.conditions).build(Lang.Branch.CONDITIONS));
+        builder.add(utils.getValueTooltip(utils, entry.functions).build(Lang.Branch.FUNCTIONS));
     }
 
-    @NotNull
-    public static TooltipBuilder getRolls(RangeValue rolls, RangeValue bonusRolls) {
-        return TooltipBuilder.value(getTotalRolls(rolls, bonusRolls).toIntString(), "x").key(Lang.Description.ROLLS);
-    }
-
-    private static RangeValue getTotalRolls(RangeValue rolls, RangeValue bonusRolls) {
-        if (bonusRolls.min() > 0 || bonusRolls.max() > 0) {
-            return bonusRolls.add(rolls);
-        } else {
-            return rolls;
-        }
+    private static void addComposite(TooltipBuilder builder, IServerUtils utils, CompositeEntryBase entry) {
+        builder.add(TooltipBuilder.array((b) -> {
+            for (LootPoolEntryContainer child : entry.children) {
+                b.add(utils.getEntryTooltip(utils, child));
+            }
+        }, Lang.Branch.ENTRIES));
+        builder.add(utils.getValueTooltip(utils, entry.conditions).build(Lang.Branch.CONDITIONS));
     }
 }