+++ ali/common/src/main/java/com/yanny/ali/plugin/server/ValueTooltipUtils.java
@@ -13,6 +13,7 @@ import net.minecraft.stats.Stat;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.storage.loot.IntRange;
+import net.minecraft.world.level.storage.loot.entries.LootPoolEntryContainer;
 import net.minecraft.world.level.storage.loot.functions.ApplyBonusCount;
 import net.minecraft.world.level.storage.loot.functions.CopyNbtFunction;
 import net.minecraft.world.level.storage.loot.functions.LootItemFunction;
@@ -31,6 +32,11 @@ public class ValueTooltipUtils {
     @NotNull
     public static TooltipBuilder getEnumTooltip(IServerUtils utils, Enum<?> condition) {
         return utils.getEnumTranslation(utils, condition);
+    }
+
+    @NotNull
+    public static TooltipBuilder getEntryTooltip(IServerUtils utils, LootPoolEntryContainer condition) {
+        return utils.getEntryTooltip(utils, condition);
     }
 
     @NotNull