+++ ali/common/src/test/java/com/yanny/ali/test/TooltipTest.java
@@ -1,20 +1,12 @@
 package com.yanny.ali.test;
 
 import com.yanny.aci.tooltip.TooltipNode;
-import com.yanny.ali.plugin.server.EntryTooltipUtils;
+import com.yanny.ali.plugin.server.TooltipUtils;
 import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.storage.loot.IntRange;
 import net.minecraft.world.level.storage.loot.entries.LootPoolSingletonContainer;
-import net.minecraft.world.level.storage.loot.functions.ApplyBonusCount;
-import net.minecraft.world.level.storage.loot.functions.LimitCount;
-import net.minecraft.world.level.storage.loot.functions.LootItemFunction;
-import net.minecraft.world.level.storage.loot.functions.LootingEnchantFunction;
-import net.minecraft.world.level.storage.loot.functions.SetItemCountFunction;
-import net.minecraft.world.level.storage.loot.predicates.BonusLevelTableCondition;
-import net.minecraft.world.level.storage.loot.predicates.ExplosionCondition;
-import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
-import net.minecraft.world.level.storage.loot.predicates.LootItemRandomChanceCondition;
-import net.minecraft.world.level.storage.loot.predicates.LootItemRandomChanceWithLootingCondition;
+import net.minecraft.world.level.storage.loot.functions.*;
+import net.minecraft.world.level.storage.loot.predicates.*;
 import net.minecraft.world.level.storage.loot.providers.number.BinomialDistributionGenerator;
 import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
 import net.minecraft.world.level.storage.loot.providers.number.UniformGenerator;
@@ -23,12 +15,12 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
+import static com.yanny.aci.test.utils.TestUtils.assertTooltip;
 import static com.yanny.ali.plugin.common.NodeUtils.getEnchantedChance;
 import static com.yanny.ali.plugin.common.NodeUtils.getEnchantedCount;
-import static com.yanny.ali.plugin.server.EntryTooltipUtils.getChanceTooltip;
-import static com.yanny.ali.plugin.server.EntryTooltipUtils.getCountTooltip;
+import static com.yanny.ali.plugin.server.TooltipUtils.getChanceTooltip;
+import static com.yanny.ali.plugin.server.TooltipUtils.getCountTooltip;
 import static com.yanny.ali.test.TooltipTestSuite.UTILS;
-import static com.yanny.aci.test.utils.TestUtils.assertTooltip;
 
 public class TooltipTest {
     @Test
@@ -246,7 +238,7 @@ public class TooltipTest {
 
     @NotNull
     private static TooltipNode itemTooltip(List<LootItemFunction> functions, List<LootItemCondition> conditions) {
-        return EntryTooltipUtils.getTooltip(UTILS, LootPoolSingletonContainer.DEFAULT_QUALITY,
+        return TooltipUtils.getTooltip(UTILS, LootPoolSingletonContainer.DEFAULT_QUALITY,
                 getEnchantedChance(UTILS, conditions, 1), getEnchantedCount(UTILS, functions), functions, conditions).build();
     }
 }